Rewrite library to parse TabDiv object for audio data
The dependency to MusicXML has been removed and the TabDiv object now
provides all the audio data. The audiolib.js and music.js libraries
have been added for realtime audio synthesis. The js code has been
rewritten to use the constructor design pattern. A pixel map has been
created to store the exact coordinates of each note to synchronize the
audio w/ the notation. The code is now much easier to follow and
audiolib.js includes a lot of great effects that will be utilized in
the future.